requirements.txt
requests_unixsocket==0.2.0
python-dotenv==0.14.0
PyYAML==5.3.1
slack-logger==0.3.1
python-telegram-handler==2.2